% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.mc}
\alias{print.mc}
\title{Print an mc object}
\usage{
\method{print}{mc}(x, ...)
}
\arguments{
\item{x}{object of class mc}

\item{...}{additional arguments passed to callies}
}
\value{
print shows a dense representation of the Monte Carlo Simulation.
}
\description{
Print function for an \code{mc} object generated by \code{\link[=future_mc]{future_mc()}}.
}
\examples{
test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 2){

data <- rnorm(n, mean = param) + x1 + x2
stat <- mean(data)
stat_2 <- var(data)

if (x2 == 5){
  stop("x2 can't be 5!")
}

return(list(mean = stat, sd = stat_2))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)




test <- future_mc(fun = test_func, repetitions = 1000, param_list = param_list)
print(test)
}
