% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.mc}
\alias{plot.mc}
\title{Plot the results of a Monte Carlo Simulation}
\usage{
\method{plot}{mc}(x, join = NULL, which_setup = NULL, parameter_comb = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{mc}, for which holds \code{simple_output = TRUE}.
See value of \code{\link[=future_mc]{future_mc()}}.}

\item{join}{A character vector containing the "nice names" for the different
parameter combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should be plotted together.
Default: Each parameter combination is plotted distinct.}

\item{which_setup}{A character vector containing the "nice names" for the different parameter
combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should be plotted.
Default: All parameter combinaitons are plotted.}

\item{parameter_comb}{test}

\item{...}{additional arguments passed to callies.}
}
\value{
returns a plot
}
\description{
Plot density plots for numeric results and bar plots for non-numeric results
of a Monte Carlo Simulation run by \code{\link[=future_mc]{future_mc()}}
}
\examples{

test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 2){

data <- rnorm(n, mean = param) + x1 + x2
stat <- mean(data)
stat_2 <- var(data)

if (x2 == 5){
  stop("x2 can't be 5!")
}

return(list(mean = stat, sd = stat_2))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)




test <- future_mc(fun = test_func, repetitions = 1000, param_list = param_list)
plot(test)
}
