% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.mc}
\alias{plot.mc}
\title{Plot the results of a Monte Carlo Simulation}
\usage{
\method{plot}{mc}(
  x,
  join = NULL,
  which_setup = NULL,
  parameter_comb = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{mc}, for which holds \code{simple_output = TRUE}.
See value of \code{\link[=future_mc]{future_mc()}}.}

\item{join}{A character vector containing the \code{nice_names} for the different
parameter combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should be plotted together.
Default: Each parameter combination is plotted distinctly.}

\item{which_setup}{A character vector containing the \code{nice_names} for the different parameter
combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should be plotted.
Default: All parameter combinations are plotted.}

\item{parameter_comb}{Alternative to \code{which_setup}. A named list whose components are named after
(some of) the parameters in \code{param_list} in \code{\link[=future_mc]{future_mc()}} and each component is a vector containing
the values for the parameters to filter by.
Default: All parameter combinations are plotted.}

\item{plot}{Boolean that specifies whether the plots should be printed while calling the function or not.
Default: \code{TRUE}}

\item{...}{ignored}
}
\value{
A list whose components are named after the outputs of \code{fun} and each component
contains an object of class \code{ggplot} and \code{gg} which can be plotted and modified with the
\link{ggplot2} functions.
}
\description{
Plot density plots for numeric results and bar plots for non-numeric results
of a Monte Carlo Simulation run by \code{\link[=future_mc]{future_mc()}}.
}
\details{
Only one of the arguments \code{join}, \code{which_setup}, and \code{paramter_comb} can be specified
at one time.
}
\examples{

test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 2){

data <- rnorm(n, mean = param) + x1 + x2
stat <- mean(data)
stat_2 <- var(data)

if (x2 == 5){
  stop("x2 can't be 5!")
}

return(list(mean = stat, sd = stat_2))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)




test <- future_mc(fun = test_func, repetitions = 1000, param_list = param_list)
returned_plot <- plot(test)

returned_plot$mean +
 ggplot2::theme_minimal() +
 ggplot2::geom_vline(xintercept = 3)



}
