% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.summary.mc}
\alias{plot.summary.mc}
\title{Plot the summarized results of a Monte Carlo Simulation}
\usage{
\method{plot}{summary.mc}(x, join = NULL, which = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{summary.mc}.}

\item{join}{A character vector containing the "nice names" for the different
parameter combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should plotted together.
Default: Each parameter combination is plotted distinct.}

\item{which}{A character vector containing the "nice names" for the different parameter
combinations (returned by \code{\link[=future_mc]{future_mc()}}), which should be plotted.
Default: All parameter combinations are plotted.}

\item{...}{additional arguments passed to callies.}
}
\value{
tet
}
\description{
Plot the summarized results of a Monte Carlo Simulation
}
\examples{
test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 2){

data <- rnorm(n, mean = param) + x1 + x2
stat <- mean(data)
stat_2 <- var(data)

if (x2 == 5){
  stop("x2 can't be 5!")
}

return(list(mean = stat, sd = stat_2))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)




test <- future_mc(fun = test_func, repetitions = 1000, param_list = param_list)
plot(summary(test))
}
