% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.mc}
\alias{summary.mc}
\title{Result summaries of a mc object}
\usage{
\method{summary}{mc}(object, sum_funcs = NULL, ...)
}
\arguments{
\item{object}{test}

\item{sum_funcs}{test}

\item{...}{test}
}
\value{
result summaries
}
\description{
Result summaries of a mc object
}
\examples{
test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 2){

data <- rnorm(n, mean = param) + x1 + x2
stat <- mean(data)
stat_2 <- var(data)

if (x2 == 5){
  stop("x2 can't be 5!")
}

return(list(mean = stat, sd = stat_2))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)




test <- future_mc(func = test_func, repetitions = 1000, param_list = param_list)
summary(test)
}
