% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_table.R
\name{tidy_mc_latex}
\alias{tidy_mc_latex}
\title{Latex table output for summary.mc object}
\usage{
tidy_mc_latex(
  x,
  repetitions_set = NULL,
  which_setup = NULL,
  which_stat = NULL,
  parameter_comb = NULL,
  caption = "Monte Carlo simulations results"
)
}
\arguments{
\item{x}{test}

\item{repetitions_set}{test}

\item{which_setup}{test}

\item{which_stat}{test}

\item{parameter_comb}{test}

\item{caption}{test}
}
\value{
return latex_table
}
\description{
Latex table output for summary.mc object
}
\examples{
test_func <- function(param = 0.1, n = 100, x1 = 1, x2 = 5, x3 = 1, x4 = 6){

  data <- rnorm(n, mean = param) + x1 + x2
  stat <- mean(data)
  stat_2 <- var(data)
  test <- LETTERS[sample(1:26, 1)]

  if(x3 == 0 & x4 == 0){
    next
  }

  if (x2 == 5){
    stop("x2 can't be 5!")
  }

  return(list(mean = stat, sd = stat_2, test = test))
}


param_list <- list(n = 10, param = seq(from = 0, to = 1, by = 0.5),
                   x1 = 1:2, x2 = 2)


test <- future_mc(fun = test_func, repetitions = 1000, param_list = param_list)

tidy_mc_latex(summary(test))
}
